// libs/features/kalite-kontrol/src/lib/kk-8115/enums/kk-8115-resource-url.ts
export enum Kk8115ResourceURL {
  QC_KAYIT_BILGILERI = '/qc-kayit-bilgileri',
  FM_PRATIGI_LIST = '/fm-pratigi-bilgileri',
  KABA_HADDELEME_LIST = '/kaba-haddeleme-paso-bilgileri',
  TESISLER_HSM = '/tesisler/hsm',
  TESISLER_GENEL = '/tesisler/genel',
  SLAB_BILGISI = '/slab-bilgisi',
  HSM_BOBIN_LIST_PATH = '/qc-onay',
  BAGIMSIZ_NUMUNE_PATH = '/bagimsiz-numune-bilgileri',
  BAGIMSIZ_MESAJ = '/bagimsiz-mesaj-bilgileri',
  YUZEY_KUSUR_KAYDI = '/yuzey-kusur-kaydi',
  KUSURLAR = '/kusurlar',
  BUTUN_AKTIF_KUSUR = '/aktif-kusur-kodu',
}